cambios generales

+++ database/migrations/2023_02_07_211845_create_tarifas_table.php
@@ -16,9 +16,9 @@ class CreateTarifasTable extends Migration
         Schema::create('tarifas', function (Blueprint $table) {
             $table->id();
             $table->string('tarifa');
-            $table->bigInteger('precio');
+            $table->float('precio');
             $table->integer('creditos');
-            $table->integer('comision');
+            $table->float('comision');
         });
     }
